+++ CHANGELOG.md
@@ -5,6 +5,16 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2023-02-22
+
+### Changed
+
+- Renamed `validation` to `form` in the `onUpdate` and `onUpdated` events.
+
+### Added
+
+- `taintedMessage` now accepts `null`.
+
 ## [0.2.0] - 2023-02-22
 
 ### Changed